feat: replace Export tab with an Options-table bulk action
Drop the dedicated Export tab from the React SPA and expose the same
export behavior as an "Export selected" button in the Options tab's bulk
action bar. Downloads a JSON file for the currently selected rows using
the existing `POST /optrion/v1/export` endpoint, so there is no change
to the wire format or to the WP-CLI export path.

Closes #48
